feat: adicionando e atualizando melhorias nos códigos de compras, usuários e ingressos

+++ src/controllers/userController.js
@@ -5,20 +5,25 @@ const { generateToken } = require('../middlewares/auth');
 exports.createUser = async (req, res) => {
     try {
       const { username, password, isAdmin } = req.body;
-  
-      if (!password) {
-        return res.status(400).json({ message: "A senha é obrigatória!" });
-      }
-  
-      const saltRounds = 10;
-      const hashedPassword = await bcrypt.hash(password, saltRounds);
-  
-      const user = await User.create({ username, password: hashedPassword, isAdmin });
-  
-      res.status(201).json(user);
+
+      const existingUser = await User.findOne({ username });
+        if (existingUser) {
+            return res.status(400).json({ message: "Nome de usuário já está em uso!" });
+        }
+
+        if (!password) {
+            return res.status(400).json({ message: "A senha é obrigatória!" });
+        }
+
+        const saltRounds = 10;
+        const hashedPassword = await bcrypt.hash(password, saltRounds);
+    
+        const user = await User.create({ username, password: hashedPassword, isAdmin });
+    
+        res.status(201).json(user);
     } catch (error) {
       console.error("Erro ao criar usuário:", error);
-      res.status(500).json({ message: "Erro ao criar usuário", error: error.message });
+      res.status(500).json({ message: "Erro ao criar usuário", error });
     }
   };
   
@@ -54,9 +59,18 @@ exports.updateUser = async (req, res) => {
 
 exports.deleteUser = async (req, res) => {
     try {
-        await User.findByIdAndDelete(req.params.id);
-        res.json({ message: 'Usuário deletado com sucesso!!!' });
+        if (req.user._id.toString() === req.params.id) {
+            return res.status(403).json({ message: "Não pode excluir sua própria conta!" });
+        }
+
+        const user = await User.findByIdAndDelete(req.params.id);
+        if (!user) {
+            return res.status(404).json({ message: "Usuário não encontrado!" });
+        }
+
+        res.status(200).json({ message: "Usuário excluído com sucesso!" });
     } catch (error) {
-        res.status(500).json({message: 'Erro ao deletar usuário', error});
+        res.status(500).json({ message: "Erro ao excluir usuário", error });
     }
 };
+
